refactor : 이미지 관련 작업 서비스의 구조 리팩토링 #53
AOP 정상 작동을 위해 외부 클래스에서 호출되도록 구조 변경
