+++ src/main/java/seondays/shareticon/image/ImageService.java
@@ -1,132 +1,36 @@
 package seondays.shareticon.image;
 
-import java.io.IOException;
-import java.time.Duration;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.retry.annotation.Backoff;
-import org.springframework.retry.annotation.Recover;
-import org.springframework.retry.annotation.Retryable;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import seondays.shareticon.exception.ImageDeleteException;
-import seondays.shareticon.exception.ImageUploadException;
-import seondays.shareticon.exception.PresignedUrlGenerationException;
 import seondays.shareticon.voucher.Voucher;
-import software.amazon.awssdk.core.exception.SdkClientException;
-import software.amazon.awssdk.core.sync.RequestBody;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
-import software.amazon.awssdk.services.s3.model.GetObjectRequest;
-import software.amazon.awssdk.services.s3.model.PutObjectRequest;
-import software.amazon.awssdk.services.s3.model.S3Exception;
-import software.amazon.awssdk.services.s3.presigner.S3Presigner;
-import software.amazon.awssdk.services.s3.presigner.model.GetObjectPresignRequest;
-import software.amazon.awssdk.services.s3.presigner.model.PresignedGetObjectRequest;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class ImageService {
 
-    private final S3Client s3Client;
-    private final S3Presigner s3Presigner;
-    @Value("${aws.s3.bucket}")
-    private String bucket;
+    private final S3ImageOperation imageOperation;
 
-    @Retryable(
-            retryFor = {SdkClientException.class},
-            noRetryFor = {S3Exception.class},
-            maxAttempts = 3,
-            backoff = @Backoff(delayExpression = "${retry.S3-Image-service.delay}", multiplier = 2)
-    )
     public String uploadImageWithRetry(VoucherImage voucherImage) {
-
-        try {
-            MultipartFile image = voucherImage.getImageFile();
-
-            String uploadTitle = makeUploadTitle("voucher", image.getOriginalFilename());
-
-            PutObjectRequest putObjectRequest = PutObjectRequest.builder()
-                    .bucket(bucket)
-                    .key(uploadTitle)
-                    .contentType(image.getContentType())
-                    .build();
-
-            RequestBody requestBody = RequestBody.fromInputStream(image.getInputStream(),
-                    image.getSize());
-
-            s3Client.putObject(putObjectRequest, requestBody);
-
-            return uploadTitle;
-        } catch (IOException e) {
-            throw new ImageUploadException();
-        }
-    }
-
-    @Recover
-    public String recoverImageUpload(Exception e, VoucherImage voucherImage) {
-        log.error("S3 이미지 업로드 시도 실패");
-        throw new ImageUploadException();
+        MultipartFile imageFile = voucherImage.getImageFile();
+        String uploadTitle = voucherImage.makeUploadTitle("test");
+        return imageOperation.uploadImageWithRetry(imageFile, uploadTitle);
     }
 
     public String getPresignedImageUrl(String objectKey, Long expirationMinutes) {
-        GetObjectRequest getObjectRequest = GetObjectRequest.builder()
-                .bucket(bucket)
-                .key(objectKey)
-                .build();
-
-        GetObjectPresignRequest presignedUrlGenerationRequest = GetObjectPresignRequest.builder()
-                .signatureDuration(Duration.ofMinutes(expirationMinutes))
-                .getObjectRequest(getObjectRequest)
-                .build();
-
-        try {
-            PresignedGetObjectRequest generatedPresignedUrl = s3Presigner.presignGetObject(
-                    presignedUrlGenerationRequest);
-            return generatedPresignedUrl.url().toString();
-        } catch (Exception e) {
-            throw new PresignedUrlGenerationException();
-        }
+        return imageOperation.getPresignedImageUrl(objectKey, expirationMinutes);
     }
 
     @Async
     public void deleteImageAsync(Voucher voucher) {
-        deleteImageWithRetry(voucher);
+        imageOperation.deleteImageWithRetry(voucher);
     }
 
-    @Retryable(
-            retryFor = {SdkClientException.class},
-            noRetryFor = {S3Exception.class},
-            maxAttempts = 3,
-            backoff = @Backoff(delayExpression = "${retry.S3-Image-service.delay}", multiplier = 2)
-    )
-    public void deleteImageWithRetry(Voucher voucher) {
-        String key = voucher.getImage();
-
-        DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
-                .bucket(bucket)
-                .key(key)
-                .build();
-
-        s3Client.deleteObject(deleteObjectRequest);
+    public void deleteImage(Voucher voucher) {
+        imageOperation.deleteImageWithRetry(voucher);
     }
 
-    @Recover
-    public void recoverImageDelete(Exception e, Voucher voucher) {
-        log.error("S3 이미지 삭제 시도 3회 실패 : {}번 쿠폰의 {}", voucher.getId(), voucher.getImage());
-        throw new ImageDeleteException();
-    }
-
-    private String makeUploadTitle(String prefix, String filename) {
-        String extension = "";
-        int index = filename.lastIndexOf('.');
-        if (index > 0) {
-            extension = filename.substring(index);
-        }
-        return prefix + "/" + UUID.randomUUID() + extension;
-    }
 }
